requirements.txt
Django==4.2.7
djangorestframework==3.14.0
cryptography==41.0.4
requests==2.31.0
python-dateutil==2.8.2
django-cors-headers==4.3.1
django-extensions==3.2.3
pytest-django==4.7.0
pytest==7.4.3
selenium==4.21.0
factory-boy==3.3.0
Faker==23.3.0
